# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Barefoot Normads",
    "description": "Barefoot Normads done by Champs in Stackup 3.",
    "contact": {}
  },
  "host": "champs-bn-api.herokuapp.com",
  "basePath": "/api/v1",
  "securityDefinitions": {},
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/trips/{tripId}": {
      "get": {
        "description": "User can get a specific trip",
        "summary": "View One trip",
        "tags": [
          "Trips"
        ],
        "operationId": "retrieveTrip",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [{
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "security token"
          },
          {
            "in": "path",
            "name": "tripId",
            "type": "integer",
            "required": true,
            "description": "parsing an trip Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {}
          },
          "401": {
            "description": "You are not allowed to access this trip",
            "headers": {}
          }
        }
      }
    }
  },
...
}